+++ app/src/main/java/com/developersbreach/composeactors/datasource/NetworkDataSource.kt
@@ -1,6 +1,8 @@
-package com.developersbreach.composeactors.repository.network
+package com.developersbreach.composeactors.datasource
 
-import com.developersbreach.composeactors.model.*
+import com.developersbreach.composeactors.data.network.JsonRemoteData
+import com.developersbreach.composeactors.data.network.RequestUrls
+import com.developersbreach.composeactors.data.model.*
 import com.developersbreach.composeactors.utils.NetworkQueryUtils
 
 /**
@@ -23,7 +25,7 @@ class NetworkDataSource {
      */
     fun getPopularActors(): List<Actor> {
         val requestUrl = requestUrls.getPopularActorsUrl()
-        val response = queryUtils.getResponseFromHttpUrl(requestUrl)
+        val response: String = queryUtils.getResponseFromHttpUrl(requestUrl)
         return jsonData.fetchActorsJsonData(response)
     }
